fixed the bugs

+++ src/main/java/school/mjc/stage0/conditions/task4/TernaryAssignment.java
@@ -3,7 +3,7 @@ package school.mjc.stage0.conditions.task4;
 public class TernaryAssignment {
     public void assignAndPrintBasedOnWhichBigger(int first, int second) {
 
-        double result = first > second ? 10 : -10;
+        int result = first > second ? 10 : -10;
         System.out.println(result);
 
     }
