+++ src/main/java/school/mjc/stage0/conditions/task4/TernaryMaxNumberPrinter.java
@@ -4,7 +4,7 @@ package school.mjc.stage0.conditions.task4;
 public class TernaryMaxNumberPrinter {
     public void printGreatest(int first, int second, int third) {
 
-        double max = first >= second ? (first >= third ? first : third) : (second >= third ? second : third);
+        int max = first >= second ? (first >= third ? first : third) : (second >= third ? second : third);
         //скобки для указания порядка действий
         System.out.println(max);
